requirements.txt
matplotlib==3.7.5
mmcv==2.1.0
mmengine==0.10.4
mmsegmentation==1.2.2
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pillow==10.3.0
scipy==1.10.1
torch==2.1.0+cu121

